Add view with five-point star
Draw a five-point star with outer corner vertices [a, b, c, d, e]. The
corner vertices are positioned in clockwise order.

__________________________________
|               (A)               |
|               * *               |
|              *   *              |
|             *     *             |
|(E)* * * * *         * * * * *(B)|
|   *                          *  |
|     *                      *    |
|       *                 *       |
|         *             *         |
|        *       *       *        |
|       *     *     *     *       |
|      *   *           *   *      |
|     * *                 * *     |
|   (D)                     (C)   |
|---------------------------------|

The vertex coordinates are expressed as a percentage of the width and
height of a bounding box.

Filling a star is possible via `percentFilled` which must be in
[0.0, 1.0]. An empty star has a `percentFilled` value of 0.0 and a
filled star has a `percentFilled` value of 1.0.