Accept JSON
Setting `allHTTPHeaderFields` on a `NSURLSessionConfiguration` object
will not pass the header fields to a data task object. Fetching
`allHTTPHeaderFields` from a `NSURLRequest` will return `nil`.

Because of this behavior, this patch refactors `GET:success:failure:` to use
`dataTaskWithRequest:completionHandler:` instead of
`dataTaskWithURL:completionHandler:` as a workaround. Setting the header fields
on a `NSURLRequest` will pass the header fields to a `NSURLRequest` object.